reorganized base code into different files
